requirements.txt
# --- Core Frameworks ---
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
fastapi-cors==0.0.6

# --- LangChain & AI (Fully compatible versions) ---
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-groq==0.1.3
langchain-huggingface==0.0.3
langsmith==0.1.58
sentence-transformers==2.7.0
faiss-cpu==1.8.0
groq==0.5.0

# --- Data Handling ---
pydantic==2.7.1
pandas==2.2.1
numpy==1.26.4
pypdf==4.2.0

# --- Utilities ---
httpx==0.27.0
requests==2.31.0

# --- Dependencies (Crucially updated versions) ---
torch==2.2.2
transformers==4.41.2      # <-- UPDATED
safetensors==0.4.3
huggingface-hub==0.23.0
tokenizers==0.19.1        # <-- This now satisfies both
scikit-learn==1.4.2
fsspec==2024.3.1